+++ lang/cs/Org.Apache.REEF.Bridge/InteropLogger.cpp
@@ -21,7 +21,7 @@
 namespace Org {
 	namespace Apache {
 		namespace REEF {
-			namespace Interop {
+			namespace Bridge {
 				/// currently not being used
 				InteropLogger::InteropLogger(JNIEnv* env, jobject jobjectInteropLogger) {
 					_env = env;
