+++ lang/cs/Org.Apache.REEF.Bridge/InteropLogger.h
@@ -27,12 +27,12 @@ using namespace System::IO;
 using namespace System::Collections::Generic;
 using namespace System::Runtime::InteropServices;
 using namespace System::Reflection;
-using namespace Org::Apache::REEF::Driver::Interop;
+using namespace Org::Apache::REEF::Driver::Bridge;
 
 namespace Org {
 	namespace Apache {
 		namespace REEF {
-			namespace Interop {
+			namespace Bridge {
 				public ref class InteropLogger : public ILogger {
 					jobject _jobjectInteropLogger;
 					jclass  _jclassInteropLogger;
